Clean up code, create modules and organize files
